A hack to allow non-integer values

+++ ajax_select/fields.py
@@ -74,7 +74,7 @@ class AutoCompleteSelectWidget(forms.widgets.TextInput):
 
         got = data.get(name, None)
         if got:
-            return long(got)
+            return got
         else:
             return None
